Refactor Position classes for simplicity, better usage.
